move XSpanIdString to lib and add Display impl
